refactor: rename label_creation.py to task_generation.py
